+++ back/src/main/java/com/rouby/user/application/UserFacade.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import com.rouby.user.application.dto.command.ResetPasswordCommand;
-import com.rouby.user.application.service.UserService;
+
 
 @Service
 @RequiredArgsConstructor
@@ -15,5 +15,4 @@ public class UserFacade {
 	public void resetPassword(Long userId, ResetPasswordCommand command) {
 		userService.resetPassword(userId, command);
 	}
-
 }
