+++ back/src/main/java/com/rouby/user/application/UserService.java
@@ -1,5 +1,6 @@
 package com.rouby.user.application;
 
+import com.rouby.user.application.dto.command.ResetPasswordCommand;
 import com.rouby.user.domain.entity.User;
 import org.springframework.stereotype.Service;
 
@@ -11,5 +12,7 @@ import org.springframework.stereotype.Service;
 
 public interface UserService {
 
-  public User findByEmail(String email);
+  User findByEmail(String email);
+
+  void resetPassword(Long userId, ResetPasswordCommand command);
 }
